Preserve bundle creation date when creating and applying delta updates (#2583)
This will preserve the file creation date of the new app bundle, but not the file creation date of any of the files inside the new app bundle because tracking those changes is complex/undesirable.

This bumps the major binary delta version to 4. A new test has been added for testing that the new bundle creation date is also preserved.